% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTable.r
\name{ReadTable}
\alias{ReadTable}
\title{Read the raw tri-axial accelerometry data csv file.}
\usage{
ReadTable(filename)
}
\arguments{
\item{filename}{The name of the csv file.}
}
\value{
The \code{ReadTable} returns a data frame with 4 columns: Index, X, Y and Z.
Index is the column for the indices of acceleration. X, Y and Z are for the acceleration
time series in each direction.
}
\description{
\code{ReadTable} reads the raw tri-axial accelerometry data in csv files.
}
\details{
The function reads csv files containing only three columns: acceleration time series
of x-, y- and z-axis.
}
\examples{
filename = system.file("extdata","sample_table.csv.gz",package="ActivityIndex")
res = ReadTable(filename)

}
