% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeScale.r
\name{TimeScale}
\alias{TimeScale}
\title{Time scale vector from 00:00:00 to 23:59:59.}
\format{A vector of characters with 86400 entries.}
\description{
A vector of length 86400 containing the time scale characters from 00:00:00 
to 23:59:59.
}
